Added `from_zero_dim_arrays_to_scalar` to numpy frontend
