+++ ivy/functional/frontends/numpy/__init__.py
@@ -229,14 +229,14 @@ numpy_scalar_to_dtype = {
     byte: ivy.int8,
     short: ivy.int16,
     intc: ivy.int32,
-    int_: ivy.int64,
     longlong: ivy.int64,
+    int_: ivy.int64,
     unsignedinteger: ivy.uint64,
     ubyte: ivy.uint8,
-    ushort: ivy.int16,
+    ushort: ivy.uint16,
     uintc: ivy.uint32,
-    uint: ivy.uint64,
     ulonglong: ivy.uint64,
+    uint: ivy.uint64,
     inexact: ivy.float64,
     floating: ivy.float64,
     half: ivy.float16,
