+++ ivy/functional/frontends/numpy/func_wrapper.py
@@ -321,3 +321,72 @@ def to_ivy_arrays_and_back(fn: Callable) -> Callable:
     and return arrays are all converted to `ndarray` instances.
     """
     return outputs_to_numpy_arrays(inputs_to_ivy_arrays(fn))
+
+
+def _is_zero_dim_array(x):
+    return x.shape == () and not ivy.isinf(x) and not ivy.isnan(x)
+
+
+def from_zero_dim_arrays_to_scalar(fn: Callable) -> Callable:
+    @functools.wraps(fn)
+    def new_fn(*args, **kwargs):
+        """
+        Calls the function, and then converts all 0 dimensional array instances in
+        the function to float numbers if out argument is not provided.
+
+        Parameters
+        ----------
+        args
+            The arguments to be passed to the function.
+
+        kwargs
+            The keyword arguments to be passed to the function.
+
+        Returns
+        -------
+            The return of the function, with 0 dimensional arrays as float numbers.
+        """
+        # call unmodified function
+        ret = fn(*args, **kwargs)
+        data = ret.data
+        if "out" in ivy.arg_names(fn):
+            # get out arg index
+            out_arg_pos = ivy.arg_info(fn, name="out")["idx"]
+            # check if out is None or out is not present in args and kwargs.
+            out_args = (
+                out_arg_pos < len(args) and args[out_arg_pos] is None
+            ) or out_arg_pos >= len(args)
+        else:
+            # no out argument accepted by the function
+            out_args = True
+
+        out_kwargs = ("out" in kwargs and kwargs["out"] is None) or "out" not in kwargs
+        if out_args and out_kwargs:
+            if isinstance(data, tuple):
+                # converting every scalar element of the tuple to float
+                data = ivy.copy_nest(data, to_mutable=True)
+                ret_idx = ivy.nested_argwhere(data, lambda x: x.shape == ())
+                try:
+                    ivy.map_nest_at_indices(
+                        data,
+                        ret_idx,
+                        lambda x: np_frontend.numpy_dtype_to_scalar[ivy.dtype(x)](x),
+                    )
+                except KeyError:
+                    raise ivy.exceptions.IvyException(
+                        "Casting to specified type is unsupported"
+                    )
+                return data
+            else:
+                # converting the scalar to float
+                if _is_zero_dim_array(data):
+                    try:
+                        return np_frontend.numpy_dtype_to_scalar[ivy.dtype(data)](data)
+                    except KeyError:
+                        raise ivy.exceptions.IvyException(
+                            f"Casting to {ivy.dtype(data)} is unsupported"
+                        )
+        return ret
+
+    new_fn.from_zero_dim_arrays_to_scalar = True
+    return new_fn
