+++ ivy_tests/test_ivy/test_misc/test_func_wrapper.py
@@ -59,11 +59,6 @@ def test_inputs_to_ivy_arrays():
     ivy.inputs_to_ivy_arrays(_fn6)(ivy.native_array(1))
 
 
-def test_from_zero_dim_arrays_to_float():
-    assert ivy.func_wrapper.from_zero_dim_arrays_to_float(_fn1)(ivy.array(1.0)) == 1.0
-    assert ivy.func_wrapper.from_zero_dim_arrays_to_float(_fn1)(ivy.array(1)) == 1.0
-
-
 def _fn7(x):
     # Assert input was converted to native array
     assert isinstance(x, ivy.NativeArray)
